Revert "Audio: Fix 64bit compilation issue of audio_stream_manager"
This reverts commit cd56f6571d70e37861569e4ab03584050d80a188.

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-4991
Reviewed-on: https://android.intel.com/343042
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>

+++ audio_stream_manager/src/StreamOut.cpp
@@ -224,7 +224,7 @@ status_t StreamOut::getPresentationPosition(uint64_t &frames, struct timespec &t
     if (!isRoutedL()) {
         return android::INVALID_OPERATION;
     }
-    uint32_t avail;
+    size_t avail;
     status_t error = getFramesAvailable(avail, timestamp);
     if (error != android::OK) {
         return error;
